Fix shadow on android context header
