+++ src/navigation/BottomTabNavigator/CustomTabBar.js
@@ -22,6 +22,8 @@ type Props = {
 
 const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
   const isDrawerOpen = useDrawerStatus() === "open";
+  const { history } = state;
+  const currentRoute = history[history.length - 1]?.key || "";
 
   const tabs = state.routes.reduce( ( tabList, route ) => {
     const { options } = descriptors[route.key];
@@ -29,8 +31,7 @@ const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
     const onPress = () => {
       navigation.navigate( { name: route.name, merge: true } );
     };
-    const { history } = state;
-    const currentRoute = history[history.length - 1]?.key || "";
+
     if ( options.meta ) {
       tabList.push(
         <View
@@ -70,6 +71,17 @@ const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
 
   const footerHeight = Platform.OS === "ios" ? "h-20" : "h-15";
 
+  // Hacky solution but is required to show ContextHeader shadow in PhotoGallery
+  // when PhotoGallery is hoisted to stack navigator, the header is rendered first
+  // and zIndex/elevation is not respected, thus the child screen cuts off the shadow
+  // there isn't a built in option to hide bottom tabs in react-navigation
+  if (
+    currentRoute.includes( "PhotoGallery" )
+    || currentRoute.includes( "GroupPhotos" )
+  ) {
+    return null;
+  }
+
   return (
     <View
       className={classNames(
@@ -82,7 +94,8 @@ const CustomTabBar = ( { state, descriptors, navigation }: Props ): Node => {
         offsetHeight: -3,
         shadowOpacity: 0.2,
         shadowRadius: 2,
-        radius: 5
+        radius: 5,
+        elevation: 5
       } )}
       accessibilityRole="tablist"
     >
